refactor(webfont-generator): split SFNT builder
