+++ packages/webfont-generator/src/sfnt/serialize.rs
