+++ packages/webfont-generator/src/svg/process.rs
@@ -2,10 +2,10 @@ use std::io::Error;
 use std::sync::Arc;
 use usvg::tiny_skia_path::Rect;
 
+use super::geometry::{bezpath_from_oxvg_path, bezpath_hash, rounded_bezpath_from_tiny_paths};
 use crate::svg::serialize::{append_path, optimize_path};
 use crate::svg::types::{ParsedGlyph, ProcessedGlyph};
 use crate::svg::winding::normalize_winding;
-use crate::ttf::{bezpath_from_oxvg_path, bezpath_hash, rounded_bezpath_from_tiny_paths};
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn process_glyph(
